src/main/java/jpabook/jpashop/repository/MemberRepositoryOld.java
package jpabook.jpashop.repository;

import jakarta.persistence.EntityManager;
import jpabook.jpashop.domain.Member;
import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Repository;

import java.util.List;

@Repository
@RequiredArgsConstructor
public class MemberRepositoryOld {

//    @PersistenceContext // @PersistenceContext를 사용하면 스프링이 EntityManager를 생성해서 주입해줌
    private final EntityManager em;

    /**
    @PersistenceUnit // EntityManagerFactory를 주입받고 싶다면, @PersistenceUnit를 사용하면 됨
    private EntityManagerFactory emf;
     */

    public void save(Member member) {
        em.persist(member);
    }

    public Member findOne(Long id) {
        return em.find(Member.class, id);
    }

    public List<Member> findAll() {
        return em.createQuery("select m from Member m", Member.class)
                .getResultList();
    }

    public List<Member> findByName(String name) {
        return em.createQuery("select m from Member m where m.name = :name", Member.class)
                .setParameter("name", name)
                .getResultList();
    }
}
